+++ src/index.ts
@@ -149,6 +149,7 @@ import * as ImportOffersFromCOATaskFactory from './task/importOffersFromCOA';
 import * as MoneyTransferTaskFactory from './task/moneyTransfer';
 import * as OnAuthorizationCreatedTaskFactory from './task/onAuthorizationCreated';
 import * as OnEventChangedTaskFactory from './task/onEventChanged';
+import * as OnOrderPaymentCompletedTaskFactory from './task/onOrderPaymentCompleted';
 import * as OnResourceUpdatedTaskFactory from './task/onResourceUpdated';
 import * as PayTaskFactory from './task/pay';
 import * as PlaceOrderTaskFactory from './task/placeOrder';
@@ -561,6 +562,7 @@ export namespace task {
         T extends TaskName.ConfirmMoneyTransfer ? ConfirmMoneyTransferTaskFactory.IData :
         T extends TaskName.OnAuthorizationCreated ? OnAuthorizationCreatedTaskFactory.IData :
         T extends TaskName.OnEventChanged ? OnEventChangedTaskFactory.IData :
+        T extends TaskName.OnOrderPaymentCompleted ? OnOrderPaymentCompletedTaskFactory.IData :
         T extends TaskName.OnResourceUpdated ? OnResourceUpdatedTaskFactory.IData :
         T extends TaskName.PlaceOrder ? PlaceOrderTaskFactory.IData :
         T extends TaskName.ConfirmRegisterService ? ConfirmRegisterServiceTaskFactory.IData :
@@ -590,6 +592,7 @@ export namespace task {
         T extends TaskName.ConfirmMoneyTransfer ? ConfirmMoneyTransferTaskFactory.IAttributes :
         T extends TaskName.OnAuthorizationCreated ? OnAuthorizationCreatedTaskFactory.IAttributes :
         T extends TaskName.OnEventChanged ? OnEventChangedTaskFactory.IAttributes :
+        T extends TaskName.OnOrderPaymentCompleted ? OnOrderPaymentCompletedTaskFactory.IAttributes :
         T extends TaskName.OnResourceUpdated ? OnResourceUpdatedTaskFactory.IAttributes :
         T extends TaskName.PlaceOrder ? PlaceOrderTaskFactory.IAttributes :
         T extends TaskName.ConfirmRegisterService ? ConfirmRegisterServiceTaskFactory.IAttributes :
@@ -619,6 +622,7 @@ export namespace task {
         T extends TaskName.ConfirmMoneyTransfer ? ConfirmMoneyTransferTaskFactory.ITask :
         T extends TaskName.OnAuthorizationCreated ? OnAuthorizationCreatedTaskFactory.ITask :
         T extends TaskName.OnEventChanged ? OnEventChangedTaskFactory.ITask :
+        T extends TaskName.OnOrderPaymentCompleted ? OnOrderPaymentCompletedTaskFactory.ITask :
         T extends TaskName.OnResourceUpdated ? OnResourceUpdatedTaskFactory.ITask :
         T extends TaskName.PlaceOrder ? PlaceOrderTaskFactory.ITask :
         T extends TaskName.ConfirmRegisterService ? ConfirmRegisterServiceTaskFactory.ITask :