+++ src/index.ts
@@ -147,6 +147,7 @@ import * as ImportEventCapacitiesFromCOATaskFactory from './task/importEventCapa
 import * as ImportEventsFromCOATaskFactory from './task/importEventsFromCOA';
 import * as ImportOffersFromCOATaskFactory from './task/importOffersFromCOA';
 import * as MoneyTransferTaskFactory from './task/moneyTransfer';
+import * as OnAssetTransactionStatusChangedTaskFactory from './task/onAssetTransactionStatusChanged';
 import * as OnAuthorizationCreatedTaskFactory from './task/onAuthorizationCreated';
 import * as OnEventChangedTaskFactory from './task/onEventChanged';
 import * as OnOrderPaymentCompletedTaskFactory from './task/onOrderPaymentCompleted';
@@ -560,6 +561,7 @@ export namespace task {
         T extends TaskName.DeleteTransaction ? DeleteTransactionTaskFactory.IData :
         T extends TaskName.GivePointAward ? GivePointAwardTaskFactory.IData :
         T extends TaskName.ConfirmMoneyTransfer ? ConfirmMoneyTransferTaskFactory.IData :
+        T extends TaskName.OnAssetTransactionStatusChanged ? OnAssetTransactionStatusChangedTaskFactory.IData :
         T extends TaskName.OnAuthorizationCreated ? OnAuthorizationCreatedTaskFactory.IData :
         T extends TaskName.OnEventChanged ? OnEventChangedTaskFactory.IData :
         T extends TaskName.OnOrderPaymentCompleted ? OnOrderPaymentCompletedTaskFactory.IData :
@@ -590,6 +592,7 @@ export namespace task {
         T extends TaskName.DeleteTransaction ? DeleteTransactionTaskFactory.IAttributes :
         T extends TaskName.GivePointAward ? GivePointAwardTaskFactory.IAttributes :
         T extends TaskName.ConfirmMoneyTransfer ? ConfirmMoneyTransferTaskFactory.IAttributes :
+        T extends TaskName.OnAssetTransactionStatusChanged ? OnAssetTransactionStatusChangedTaskFactory.IAttributes :
         T extends TaskName.OnAuthorizationCreated ? OnAuthorizationCreatedTaskFactory.IAttributes :
         T extends TaskName.OnEventChanged ? OnEventChangedTaskFactory.IAttributes :
         T extends TaskName.OnOrderPaymentCompleted ? OnOrderPaymentCompletedTaskFactory.IAttributes :
@@ -620,6 +623,7 @@ export namespace task {
         T extends TaskName.DeleteTransaction ? DeleteTransactionTaskFactory.ITask :
         T extends TaskName.GivePointAward ? GivePointAwardTaskFactory.ITask :
         T extends TaskName.ConfirmMoneyTransfer ? ConfirmMoneyTransferTaskFactory.ITask :
+        T extends TaskName.OnAssetTransactionStatusChanged ? OnAssetTransactionStatusChangedTaskFactory.ITask :
         T extends TaskName.OnAuthorizationCreated ? OnAuthorizationCreatedTaskFactory.ITask :
         T extends TaskName.OnEventChanged ? OnEventChangedTaskFactory.ITask :
         T extends TaskName.OnOrderPaymentCompleted ? OnOrderPaymentCompletedTaskFactory.ITask :