+++ src/taskName.ts
@@ -49,6 +49,7 @@ export enum TaskName {
     ImportEventsFromCOA = 'importEventsFromCOA',
     ImportOffersFromCOA = 'importOffersFromCOA',
     MoneyTransfer = 'moneyTransfer',
+    OnAssetTransactionStatusChanged = 'onAssetTransactionStatusChanged',
     OnAuthorizationCreated = 'onAuthorizationCreated',
     OnEventChanged = 'onEventChanged',
     OnResourceUpdated = 'onResourceUpdated',