+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 資産取引検索条件拡張
 - 決済取引確定タスクにprocessOrderオプションを追加
 - タスク検索条件拡張
+- onOrderPaymentCompletedタスクを追加
 
 ### Changed