+++ CHANGELOG.md
@@ -20,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - 決済方法区分未指定の対面決済承認に対応
+- 決済取引のobjectを最適化(paymentMethodId,paymentMethod.identifier)
 
 ### Deprecated