+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - onAssetTransactionStatusChangedタスクを追加
 - 注文にinvoice.paymentMethod.identifierを決済方法区分コードとして保証
 - 決済取引確定時の決済方法区分指定を追加
+- 注文にpreviousOrderStatusを追加
 
 ### Changed